[Modify]#13 BaseFragment 수정
 - BindingFragment, BaseFragmentUseHandler, DisposableManageFragment 계층 구조로 변경
 - 범용적으로 사용 가능한 함수들은 extension function으로 변경
 - Progress, BaseViewModel 등에 대한 기능 제거
